feat: define partition key type and sort key type with a string instead of an enum
Define partition key type and sort key type with a string instead of an enum. This makes it easier
to load configuration from a simple json object like a json file.

BREAKING CHANGE: Define partition key type and sort key type with a string instead of an enum. This
makes it easier to load configuration from a simple json object like a json file.